+++ src/data/midi/mod.rs
@@ -1,3 +1,0 @@
-//! MIDI message and related types.
-
-pub mod message;