+++ src/data/mod.rs
@@ -1,4 +1,3 @@
 //! Primitives and structures used in USB MIDI data.
 
 pub mod byte;
-pub mod midi;